+++ main.ts
@@ -1,7 +1,7 @@
 import {
   archiveCurrentS4SceneCollections,
 } from "./src/obsSceneCollectionManagement.ts";
-import { createDirectories } from "./src/directories.ts";
+import { createDirectories } from "./src/folderManagement.ts";
 
 import { generateSceneCollectionFromWhiteboard } from "./src/obsSceneCollectionManagement.ts";
 import { copySceneCollectionAssets } from "./src/assetsManagement.ts";
