+++ src/assetsManagement.ts
@@ -3,7 +3,7 @@ import { copy } from "https://deno.land/std@0.224.0/fs/copy.ts";
 import {
   localSceneCollectionAssetsFolder,
   s4vrSceneCollectionAssetsFolder,
-} from "./directories.ts";
+} from "./folderManagement.ts";
 
 export async function copySceneCollectionAssets() {
   await copy(
