+++ src/folderManagement.ts
@@ -1,0 +1,42 @@
+import * as path from "https://deno.land/std/path/mod.ts";
+
+export const sceneCollectionPrefix = "Sunday_Service";
+
+export async function createDirectories() {
+  await Deno.mkdir(
+    s4vrSceneCollectionArchiveFolder,
+    { recursive: true },
+  );
+  await Deno.mkdir(
+    s4vrPreRecordsFolder,
+    { recursive: true },
+  );
+}
+
+// TODO:  I think this is the only thing blocking mac compatibility,
+// and I know at least one person streams from a mac.
+export const obsSceneCollectionFolder = path.join(
+  Deno.env.get("APPDATA") ??
+    "~/Library/Application Support/obs-studio/basic/scenes",
+  "\\obs-studio\\basic\\scenes",
+);
+
+export const s4vrSceneCollectionArchiveFolder = path.join(
+  Deno.env.get("APPDATA") ??
+    "~/Library/Application Support/obs-studio/basic/scenes",
+  "\\s4vr\\archives\\sceneCollections",
+);
+
+export const s4vrPreRecordsFolder = path.join(
+  Deno.env.get("APPDATA") ??
+    "~/Library/Application Support/obs-studio/basic/scenes",
+  "\\s4vr\\prerecords",
+);
+
+export const s4vrSceneCollectionAssetsFolder = path.join(
+  Deno.env.get("APPDATA") ??
+    "~/Library/Application Support/obs-studio/basic/scenes",
+  "\\s4vr\\sceneColletions\\assets",
+);
+
+export const localSceneCollectionAssetsFolder = "./sceneCollectionAssets";
