+++ src/whiteboard.ts
@@ -1,20 +1,9 @@
-export async function getWhiteboardSlotsMapped() {
-  return (await getWhiteboard()).event.slots.map(mapWhiteboardSlotToObsSlot);
-}
-
-async function getWhiteboard() {
+export async function getWhiteboard() {
   const response = await fetch("https://whiteboard-diczrrhb6a-uc.a.run.app/");
   const whiteboard = await response.json();
+
+  console.log(`Whiteboard State`);
+  console.log(JSON.stringify(whiteboard, null, 4));
+
   return whiteboard;
-}
-
-function mapWhiteboardSlotToObsSlot(slot: any) {
-  const returnSlot = { ...slot };
-
-  if (returnSlot.slotType === "TWITCH") {
-    returnSlot.mediaSourceUrl =
-      `https://twitch.tv/${slot.twitchUserName}/embed?frameborder="0"`;
-  }
-
-  return returnSlot;
 }
